requirements.txt
attrs==23.1.0
bcrypt==4.1.2
cads-api-client==1.1.0
cdsapi==0.7.0
cffi==1.16.0
cfgrib==0.9.13.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-yasg==1.21.7
eccodes==1.7.1
et-xmlfile==1.1.0
Faker==16.1.0
findlibs==0.0.5
idna==3.7
inflection==0.5.1
jsonschema==4.17.3
multiurl==0.3.1
mysqlclient==2.2.0
numpy==2.0.0
openpyxl==3.1.5
packaging==23.2
pandas==2.2.2
paramiko==3.4.0
Pillow==10.0.1
psycopg2==2.9.6
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.32.3
six==1.16.0
sshtunnel==0.4.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wincertstore==0.2
xarray==2024.6.0
